astar, tictactoe: made the compiler behave more pedantically (`-pedantic`) and show all warnings (`-Wall`)

+++ 2/main.cpp
@@ -24,7 +24,7 @@ int main()
 
     SlidingBlocks::State begin(std::move(beginMatrix), emptyCellY, emptyCellX), end(size, 1);
     auto path = std::move(SlidingBlocks::FindPath(begin, end));
-    printf("%d\n", path.size() - 1);
+    printf("%d\n", (int)path.size() - 1);
 
     for (auto node: path)
         switch (node->GetMovement())
